Updated the timescheme documentation
