+++ docs/sections/api/specfem/index.rst
@@ -16,6 +16,7 @@
     program/index
     receivers/index
     sources/index
+    timescheme/index
     data_access/index
     source_time_functions/index
     chunk_edge/index
