+++ docs/sections/api/specfem/timescheme/index.rst
@@ -1,0 +1,12 @@
+``specfem::time_scheme``
+========================
+
+.. doxygennamespace:: specfem::time_scheme
+    :desc-only:
+
+
+.. toctree::
+    :maxdepth: 1
+
+    timescheme
+    newmark
