+++ docs/sections/api/specfem/timescheme/newmark_combined.rst
@@ -1,0 +1,8 @@
+
+.. _newmark_combined:
+
+Newmark Combined Time Scheme
+============================
+
+.. doxygenclass:: specfem::time_scheme::newmark< AssemblyFields, specfem::simulation::type::combined >
+    :members:
