+++ docs/sections/api/specfem/timescheme/newmark_forward.rst
@@ -1,0 +1,9 @@
+
+
+.. _newmark_forward:
+
+Newmark Forward Time Scheme
+===========================
+
+.. doxygenclass:: specfem::time_scheme::newmark< AssemblyFields, specfem::simulation::type::forward >
+    :members:
