+++ docs/sections/api/specfem/timescheme/timescheme.rst
@@ -1,0 +1,5 @@
+``specfem::time_scheme::time_scheme``
+=====================================
+
+.. doxygenclass:: specfem::time_scheme::time_scheme
+    :members:
