--- docs/sections/api/timescheme/index.rst
@@ -1,18 +1,0 @@
-
-.. _timescheme:
-
-Time Scheme API
-================
-
-The ``TimeScheme`` class provides methods to compute the time evolution of a wavefield.
-
-.. doxygenclass:: specfem::time_scheme::time_scheme
-   :members:
-
-Types of Time Schemes
----------------------
-
-.. toctree::
-    :maxdepth: 1
-
-    newmark
