--- docs/sections/api/timescheme/newmark.rst
@@ -1,17 +1,0 @@
-
-.. _timescheme_newmark:
-
-Newmark Time Scheme
-====================
-
-.. doxygenclass:: specfem::time_scheme::newmark
-    :members:
-
-Implementation Details
-----------------------
-
-.. toctree::
-    :maxdepth: 1
-
-    newmark_forward
-    newmark_combined
