--- docs/sections/api/timescheme/newmark_combined.rst
@@ -1,8 +1,0 @@
-
-.. _newmark_combined:
-
-Newmark Combined Time Scheme
-============================
-
-.. doxygenclass:: specfem::time_scheme::newmark< specfem::simulation::type::combined >
-    :members:
