--- docs/sections/api/timescheme/newmark_forward.rst
@@ -1,9 +1,0 @@
-
-
-.. _newmark_forward:
-
-Newmark Forward Time Scheme
-===========================
-
-.. doxygenclass:: specfem::time_scheme::newmark< specfem::simulation::type::forward >
-    :members:
